refactor(ledger): split the CRM route file into per-resource route files :recycle:
CRM is folded into the ledger binary, so its routes had no reason to keep a
separate cross-resource registrar while every ledger resource owns a
<resource>_routes.go. crm_routes.go is gone and its surface is redistributed
onto five per-resource files that follow the ledger pattern exactly — a
Register<R>Routes Huma registrar, a Register<R>V2RoutesToApp mount and a
register<R>RoutesToApp body:

- holder_routes.go, holder_accounts_routes.go, instrument_routes.go,
  encryption_routes.go, audit_routes.go

The Huma registrars move out of the handler files to sit beside their Fiber
chain, matching account_handler.go carrying no registrar. The three conditional
seams keep their guards, now as an early return in their own
Register<R>V2RoutesToApp rather than an if inside a shared body: holder-accounts
when no ledger account-query backing is wired, encryption and audit outside
envelope mode. Holder and instrument stay unguarded, as before.

crmOpSuffixV2 moves to routes_op_suffix.go next to the core-ledger suffixes it
is deliberately separate from. ApplicationName is retired: it duplicated
routes.go's midazName, and the name also meant "ledger" in bootstrap. Its X1
RBAC lock moves to routes_test.go as TestMidazAppName_IsTheX1AuthzNamespace, and
the nil-guard suite moves to huma_contract_mount_test.go, which is where mount
composition is tested.

Route order, authz tuples, tenant options, operation IDs and paths are
unchanged: TestContractSpecMatchesRoutes and TestCRMCompositionSurfaceMountedOnGroup
pass and no OpenAPI golden moved. Pointers in permissions.yaml, the X1 RBAC
runbook, RBAC-NAMESPACES, crm-field-encryption, STRUCTURE, AGENTS and the llms
manifests follow the split.
