+++ components/ledger/internal/adapters/http/in/audit_handler.go
@@ -18,7 +18,7 @@ import (
 // It mirrors the asset exemplar (asset_handler.go); see that file's header
 // for the full conventions. Audit-specific notes:
 //
-//  1. AUTH is appName "midaz" (crm_routes.go ApplicationName), resource
+//  1. AUTH is appName "midaz" (routes.go midazName), resource
 //     "protection". The Fiber guard chain is Bearer-only, so the per-op Security
 //     metadata here is Bearer-only too — SPEC metadata only;
 //     runtime auth stays the Fiber guard chain (auth.Authorize("midaz","protection",
@@ -89,24 +89,4 @@ func (handler *AuditHandler) GetAuditEvents(ctx context.Context, in *GetAuditEve
 	}
 
 	return &GetAuditEventsResponse{Status: http.StatusOK, Body: envelope}, nil
-}
-
-// RegisterAuditRoutes registers the protection-audit operation on the
-// given Huma API. It is the per-file seam the unified server calls (conditionally,
-// only in envelope encryption mode — mirroring the Fiber `if auditHandler != nil`
-// guard in crm_routes.go); the auth ("midaz","protection","get") + tenant +
-// ParseUUIDPathParameters("organization") middleware chain is attached on the
-// versioned Fiber group BEFORE the Huma terminal, not here. Paths are GROUP-RELATIVE
-// (see asset_handler.go's RegisterAssetRoutes header for the rationale).
-//
-// opSuffix is appended to the operation ID — see crmOpSuffixV2.
-func RegisterAuditRoutes(api huma.API, h *AuditHandler, opSuffix string) {
-	huma.Register(api, huma.Operation{
-		OperationID: "getAuditEvents" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        "/organizations/{organization_id}/protection/audit",
-		Summary:     "List Protection Audit Events",
-		Tags:        []string{"Protection"},
-		Security:    secAuditBearer,
-	}, h.GetAuditEvents)
 }
