+++ components/ledger/internal/adapters/http/in/huma_contract_mount.go
@@ -203,7 +203,11 @@ func (d HumaMountDeps) MountV2(group fiber.Router, api huma.API) {
 	RegisterCountTransactionV2RoutesToApp(group, api, d.Auth, d.Transaction, d.TransactionOptions)
 	RegisterBalanceV2RoutesToApp(group, api, d.Auth, d.Balance, d.TransactionOptions)
 	RegisterOperationV2RoutesToApp(group, api, d.Auth, d.Operation, d.TransactionOptions)
-	RegisterCRMV2RoutesToApp(group, api, d.Auth, d.Holder, d.Instrument, d.HolderAccounts, d.Encryption, d.Audit, d.CRMOptions)
+	RegisterHolderV2RoutesToApp(group, api, d.Auth, d.Holder, d.CRMOptions)
+	RegisterHolderAccountsV2RoutesToApp(group, api, d.Auth, d.HolderAccounts, d.CRMOptions)
+	RegisterInstrumentV2RoutesToApp(group, api, d.Auth, d.Instrument, d.CRMOptions)
+	RegisterEncryptionV2RoutesToApp(group, api, d.Auth, d.Encryption, d.CRMOptions)
+	RegisterAuditV2RoutesToApp(group, api, d.Auth, d.Audit, d.CRMOptions)
 	RegisterFeesV2RoutesToApp(group, api, d.Auth, d.FeePackage, d.Fee, d.BillingPackage, d.BillingCalculate, d.FeesOptions)
 	RegisterCompositionV2RoutesToApp(group, api, d.Auth, d.Composition, d.CompositionOptions)
 	RegisterOperationRouteV2RoutesToApp(group, api, d.Auth, d.OperationRoute, d.TransactionOptions)
